perf(mem): Step-4 detection loads frames as float32 too
Detection loaded each frame as float64 (488 MB) then downcast to float32 for the
SEP path — the float64 intermediate is pure waste. Load float32 directly (244 MB;
the SEP astype is now a no-op with copy=False). Same rationale as the Step-7 fix;
together with the 2-worker caps this keeps the whole reprocess under ~1-1.5 GB.

+++ apex/analysis/detection.py
@@ -243,9 +243,11 @@ def run_detection(file_list, params, data_dir, cache_dir, use_cropped=False,
                     detect_engine=getattr(P, 'detect_engine', 'segm'),
                 )
 
-                # Load FITS
+                # Load FITS. float32, not float64: halves per-frame RAM (244 MB
+                # vs 488 MB for a 61 MP frame) and the SEP path downcasts to
+                # float32 anyway, so this also drops the float64 intermediate.
                 with fits.open(file_path) as hdul:
-                    data = hdul[0].data.astype(float)
+                    data = hdul[0].data.astype(np.float32)
                     header = hdul[0].header
 
                 # Get filter - preserve original case from header
